+++ packages/sampler/stories/qa-stories/ExpandableList.js
@@ -1,5 +1,6 @@
 import ExpandableList from '@enact/moonstone/ExpandableList';
 import Button from '@enact/moonstone/Button';
+import {RadioControllerDecorator} from '@enact/ui/RadioDecorator';
 import React from 'react';
 import {storiesOf, action} from '@kadira/storybook';
 import {withKnobs, boolean, text, select} from '@kadira/storybook-addon-knobs';
@@ -34,14 +35,20 @@ class ExpandableListChildrenLengthUpdate extends React.Component {
 	}
 }
 
+const ExpandableGroup = RadioControllerDecorator('div');
+ExpandableGroup.displayName = 'ExpandableGroup';
+ExpandableList.displayName = 'ExpandableList';
+
 storiesOf('ExpandableList')
 	.addDecorator(withKnobs)
 	.addWithInfo(
 		'with children length update',
 		() => (
 			<ExpandableListChildrenLengthUpdate
-				autoClose={boolean('autoClose', false)}
+				closeOnSelect={boolean('closeOnSelect', false)}
 				disabled={boolean('disabled', false)}
+				noAutoClose={boolean('noAutoClose', false)}
+				noLockBottom={boolean('noLockBottom', false)}
 				noneText={text('noneText', 'nothing selected')}
 				onSelect={action('onSelect')}
 				onClose={action('onClose')}
@@ -50,5 +57,37 @@ storiesOf('ExpandableList')
 				select={select('select', ['single', 'radio', 'multiple'], 'single')}
 				title={text('title', 'title')}
 			/>
+		)
+	)
+	.addWithInfo(
+		'grouped',
+		() => (
+			<ExpandableGroup>
+				<ExpandableList title="First">
+					{['One', 'Two', 'Three']}
+				</ExpandableList>
+				<ExpandableList title="Second">
+					{['Fourth', 'Fifth', 'Sixth']}
+				</ExpandableList>
+				<ExpandableList title="Third">
+					{['Seventh', 'Eighth', 'Ninth']}
+				</ExpandableList>
+			</ExpandableGroup>
+		)
+	)
+	.addWithInfo(
+		'with multiples (to test "lockBottom" prop)',
+		() => (
+			<div>
+				<ExpandableList title="First">
+					{['One', 'Two', 'Three']}
+				</ExpandableList>
+				<ExpandableList title="Second">
+					{['Fourth', 'Fifth', 'Sixth']}
+				</ExpandableList>
+				<ExpandableList title="Third">
+					{['Seventh', 'Eighth', 'Ninth']}
+				</ExpandableList>
+			</div>
 		)
 	);