+++ packages/spotlight/src/container.js
@@ -1,6 +1,6 @@
 import {forward} from '@enact/core/handle';
 import hoc from '@enact/core/hoc';
-import React from 'react';
+import React, {PropTypes} from 'react';
 
 import Spotlight from './spotlight';
 import {spottableClass} from './spottable';
@@ -30,13 +30,13 @@ const defaultConfig = {
 	enterTo: 'last-focused',
 
 	/**
-	 * Restricts or prioritizes navigation when focus attempts to leave the container.
+	 * Whether the container will preserve the id when it unmounts.
 	 *
-	 * @type {String}
-	 * @default 'none'
+	 * @type {Boolean}
+	 * @default false
 	 * @public
 	 */
-	restrict: 'none'
+	preserveId: false
 };
 
 /**
@@ -45,7 +45,7 @@ const defaultConfig = {
  *
  * @example
  *	const DefaultContainer = SpotlightContainerDecorator(Component);
- *	const SelfRestrictedContainer = SpotlightContainerDecorator({restrict: 'self-only'}, Component);
+ *	const FocusDefaultContainer = SpotlightContainerDecorator({enterTo: 'default-element'}, Component);
  *
  * To specify a default element to spot in a container, utilize the `spotlightDefaultClass`.
  *
@@ -69,14 +69,49 @@ const defaultConfig = {
 const SpotlightContainerDecorator = hoc(defaultConfig, (config, Wrapped) => {
 	const forwardMouseEnter = forward(enterEvent);
 	const forwardMouseLeave = forward(leaveEvent);
+	const {preserveId} = config;
 
 	return class extends React.Component {
-		static displayName = 'SpotlightContainerDecorator'
+		static displayName = 'SpotlightContainerDecorator';
+
+		static propTypes = /** @lends spotlight/SpotlightContainerDecorator.SpotlightContainerDecorator */ {
+			/**
+			 * Specifies the container id. If the value is `null`, an id will be generated.
+			 *
+			 * @type {String}
+			 * @public
+			 */
+			containerId: PropTypes.string,
+
+			/**
+			 * Whether or not the container is in muted mode.
+			 *
+			 * @type {Boolean}
+			 * @default false
+			 * @public
+			 */
+			spotlightMuted: PropTypes.bool,
+
+			/**
+			 * Restricts or prioritizes navigation when focus attempts to leave the container. It
+			 * can be either 'none', 'self-first', or 'self-only'.
+			 *
+			 * @type {String}
+			 * @default 'none'
+			 * @public
+			 */
+			spotlightRestrict: PropTypes.oneOf(['none', 'self-first', 'self-only'])
+		}
+
+		static defaultProps = {
+			spotlightMuted: false,
+			spotlightRestrict: 'none'
+		}
 
 		constructor (props) {
 			super(props);
 			this.state = {
-				containerId: ''
+				id: this.props.containerId || Spotlight.add()
 			};
 		}
 
@@ -85,7 +120,7 @@ const SpotlightContainerDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			while (elem && elem !== document && elem.nodeType === 1) {
 				containerId = elem.getAttribute('data-container-id');
 				if (containerId &&
-						containerId !== this.state.containerId &&
+						containerId !== this.state.id &&
 						elem.getAttribute('data-container-disabled') === 'true') {
 
 					return false;
@@ -94,21 +129,37 @@ const SpotlightContainerDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			}
 		}
 
-		componentWillMount () {
-			const containerId = Spotlight.add(),
-				selector = '[data-container-id="' + containerId + '"]:not([data-container-disabled="true"]) .' + spottableClass,
-				cfg = Object.assign({}, config, {selector, navigableFilter: this.navigableFilter});
+		componentWillReceiveProps (nextProps) {
+			if (this.props.containerId !== nextProps.containerId) {
+				Spotlight.remove(this.props.containerId);
+				Spotlight.add(nextProps.containerId);
+				this.setState({
+					id: nextProps.containerId
+				});
+			}
+		}
 
-			Spotlight.set(containerId, cfg);
-			this.setState({containerId: containerId});
+		componentDidUpdate (prevProps) {
+			if (this.props.spotlightRestrict !== prevProps.spotlightRestrict) {
+				Spotlight.set(this.state.id, {restrict: this.props.spotlightRestrict});
+			}
+		}
+
+		componentWillMount () {
+			const selector = '[data-container-id="' + this.state.id + '"]:not([data-container-disabled="true"]) .' + spottableClass,
+				cfg = Object.assign({}, config, {selector, navigableFilter: this.navigableFilter, restrict: this.props.spotlightRestrict});
+
+			Spotlight.set(this.state.id, cfg);
 		}
 
 		componentWillUnmount () {
-			Spotlight.remove(this.state.containerId);
+			if (!preserveId) {
+				Spotlight.remove(this.state.id);
+			}
 		}
 
 		handleMouseEnter = (ev) => {
-			Spotlight.setActiveContainer(this.state.containerId);
+			Spotlight.setActiveContainer(this.state.id);
 			forwardMouseEnter(ev, this.props);
 		}
 
@@ -118,14 +169,19 @@ const SpotlightContainerDecorator = hoc(defaultConfig, (config, Wrapped) => {
 		}
 
 		render () {
-			const containerId = this.state.containerId,
-				props = Object.assign({}, this.props);
+			const {spotlightMuted, ...rest} = this.props;
+			delete rest.containerId;
+			delete rest.spotlightRestrict;
 
-			props['data-container-id'] = containerId;
-			props[enterEvent] = this.handleMouseEnter;
-			props[leaveEvent] = this.handleMouseLeave;
+			rest['data-container-id'] = this.state.id;
+			rest[enterEvent] = this.handleMouseEnter;
+			rest[leaveEvent] = this.handleMouseLeave;
 
-			return <Wrapped {...props} />;
+			if (spotlightMuted) {
+				rest['data-container-muted'] = spotlightMuted;
+			}
+
+			return <Wrapped {...rest} />;
 		}
 	};
 });